requirements.txt
streamlit
azure-ai-documentintelligence
azure-identity
azure-keyvault-secrets
python-dotenv